New Stuff

+++ Unity RTS/Assets/Scripts/UnitStats.cs
@@ -6,7 +6,7 @@ public abstract class UnitStats : MonoBehaviour
 {
     //TODO: Make enum and assign a 0 or 1 for friendly or not (Idea for now)
     public abstract void Die();
-    public abstract void TakeDamage(int damage);
+    public abstract void TakeDamage(float damage);
     public abstract void Fire();
 
 }
